requirements.txt
fastapi~=0.115.6
pydantic~=2.10.3
uvicorn~=0.32.1
gunicorn~=23.0.0
pytesseract~=0.3.13
pillow~=11.1.0
python-logstash-async~=3.0.0